feat: check working tree is clean
