+++ package.json
@@ -76,9 +76,12 @@
     "chalk": "^5.3.0",
     "esbuild": "^0.18.11",
     "eslint": "^8.41.0",
+    "execa": "^7.1.1",
     "figlet": "^1.6.0",
     "glob": "^8.1.0",
     "mocha": "^10.2.0",
+    "ora": "^6.3.1",
+    "semver": "^7.5.4",
     "vsce": "^2.15.0"
   },
   "dependencies": {
