+++ scripts/release/interfaces.ts
@@ -1,0 +1,10 @@
+/**
+ * CLI exit codes.
+ *
+ * @see https://nodejs.org/api/process.html#process_exit_codes
+ */
+export enum ExitCode {
+  Success = 0,
+  FatalError = 1,
+  InvalidArgument = 9,
+}
